refactor: 상수의 용도별로 파일 분리
