+++ src/constants/database.ts
@@ -1,0 +1,3 @@
+export const STORE = {
+  keyPrefix: 'lunch_',
+};
